+++ spec/elastic_apm/stacktrace_builder_spec.rb
@@ -22,7 +22,7 @@ module ElasticAPM
           expect(last_frame.pre_context.last).to match(/def actual_exception/)
           expect(last_frame.context_line).to match(%r{1 / 0})
           expect(last_frame.post_context.first).to match(/rescue/)
-          expect(last_frame.filename).to eq 'spec_helper.rb'
+          expect(last_frame.filename).to eq 'support/exception_helpers.rb'
         end
       end